+++ packages/react-langgraph/src/useLangGraphRuntime.ts
@@ -139,10 +139,25 @@ export const useLangGraphRuntime = ({
         feedback?: FeedbackAdapter;
       }
     | undefined;
+  /**
+   * Event handlers for various LangGraph stream events
+   */
   eventHandlers?: {
+    /**
+     * Called when metadata is received from the LangGraph stream
+     */
     onMetadata?: OnMetadataEventCallback;
+    /**
+     * Called when informational messages are received from the LangGraph stream
+     */
     onInfo?: OnInfoEventCallback;
+    /**
+     * Called when errors occur during LangGraph stream processing
+     */
     onError?: OnErrorEventCallback;
+    /**
+     * Called when custom events are received from the LangGraph stream
+     */
     onCustomEvent?: OnCustomEventCallback;
   } | undefined;
 }) => {
